+++ demos/backend/src/minimal.ts
@@ -47,6 +47,6 @@ const server = Server.configure({
   // },
 })
 
-server.enableMessageLogging()
+// server.enableMessageLogging()
 
 server.listen()