+++ packages/extension-rocksdb/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.82](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-rocksdb@1.0.0-alpha.81...@hocuspocus/extension-rocksdb@1.0.0-alpha.82) (2021-11-26)
+
+**Note:** Version bump only for package @hocuspocus/extension-rocksdb
+
+
+
+
+
+# [1.0.0-alpha.81](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-rocksdb@1.0.0-alpha.80...@hocuspocus/extension-rocksdb@1.0.0-alpha.81) (2021-11-24)
+
+**Note:** Version bump only for package @hocuspocus/extension-rocksdb
+
+
+
+
+
+# [1.0.0-alpha.80](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-rocksdb@1.0.0-alpha.79...@hocuspocus/extension-rocksdb@1.0.0-alpha.80) (2021-11-22)
+
+**Note:** Version bump only for package @hocuspocus/extension-rocksdb
+
+
+
+
+
 # [1.0.0-alpha.79](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-rocksdb@1.0.0-alpha.78...@hocuspocus/extension-rocksdb@1.0.0-alpha.79) (2021-11-10)
 
 **Note:** Version bump only for package @hocuspocus/extension-rocksdb