+++ packages/extension-throttle/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.37](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-throttle@1.0.0-alpha.36...@hocuspocus/extension-throttle@1.0.0-alpha.37) (2021-11-26)
+
+**Note:** Version bump only for package @hocuspocus/extension-throttle
+
+
+
+
+
+# [1.0.0-alpha.36](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-throttle@1.0.0-alpha.35...@hocuspocus/extension-throttle@1.0.0-alpha.36) (2021-11-24)
+
+**Note:** Version bump only for package @hocuspocus/extension-throttle
+
+
+
+
+
+# [1.0.0-alpha.35](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-throttle@1.0.0-alpha.34...@hocuspocus/extension-throttle@1.0.0-alpha.35) (2021-11-22)
+
+**Note:** Version bump only for package @hocuspocus/extension-throttle
+
+
+
+
+
 # [1.0.0-alpha.34](https://github.com/ueberdosis/hocuspocus/compare/@hocuspocus/extension-throttle@1.0.0-alpha.33...@hocuspocus/extension-throttle@1.0.0-alpha.34) (2021-11-10)
 
 **Note:** Version bump only for package @hocuspocus/extension-throttle